% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_kapita.R
\name{colors_kapita}
\alias{colors_kapita}
\title{Get kapita Brand Colors}
\usage{
colors_kapita()
}
\value{
A list with kapita brand colors
}
\description{
Get kapita Brand Colors
}
\examples{
colors_kapita()
}
