% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_company_colors.R
\name{get_company_colors}
\alias{get_company_colors}
\title{get company colors}
\usage{
get_company_colors(company = "kapita")
}
\arguments{
\item{company}{A string for the company that the colors are needed for}
}
\value{
A list of brand colors of the company
}
\description{
get company colors
}
\examples{
get_company_colors(company = "kapita")
}
