% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_kapita.R
\name{theme_kapita}
\alias{theme_kapita}
\title{Get the ggplot theme used for kapita}
\usage{
theme_kapita()
}
\value{
A ggplot2 theme element used for kapita plots
}
\description{
Get the ggplot theme used for kapita
}
\examples{
theme_kapita()
}
